move PythnCodeAnalyzer to dedicated file
